refactor for java standard 1 class per file.
